% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{dataDTWUMI}
\alias{dataDTWUMI}
\title{A multivariate times series consisting of three signals as example for DTWUMI package}
\format{A list storing two data frames with three columns each.
The first table contains the original complete simulated data.
The second table contains the same simulated data with one large gap added within each signal.}
\usage{
dataDTWUMI
}
\description{
A multivariate times series consisting of three signals as example for DTWUMI package
}
\keyword{datasets}
