% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_3_1Imputing_1NA.R
\name{imp_1NA}
\alias{imp_1NA}
\title{Imputing gaps of size 1}
\usage{
imp_1NA(data, pos1)
}
\arguments{
\item{data}{a univariate signal}

\item{pos1}{the position of the begining of gaps of size 1, as obtained using Indexes_size_missing_multi() function}
}
\value{
returns a new vector of same size with imputed values
}
\description{
Imputes isolated missing values based on the average of nearest neighbours.
}
\author{
DEZECACHE Camille, PHAN Thi Thu Hong, POISSON-CAILLAULT Emilie
}
